.transitions() method created
This method shall display the transitions of the automata under a transitions table, making it more readable. So far, however, only the header has been implemented in the finiteAutomaton class, working for all currently-implemented automata types and adjusting the width of the columns automatically, based on the maximum number of destinations possible to arrive from the same origin, considering the same symbol.

Some logic is necessary to get this right, however: since there are situations in which there are no transitions from/to every state in the automaton, the empty symbol must be displayed. Once this logic is set, the method will have the flexibility needed to display every possible situation in a readable manner, making it possible even to integrate it under the finiteAutomaton's .__repr__() method.

Planning, planning, planning.