.process_symbol() improvements; NFAE better supported
Thanks to adaptations in the .process_symbol() method, which aim to provide superior flexibility, and to improvements under the .process_word() method to support empty moves, the finiteAutomaton class is better suited for automatons with empty moves.

Further adaptations made the function work without raising Exceptions, aiming to provide a seamless experience instead of complaining about every problem. Everything is explained in the verbose mode.