dfpa.py created
It works flawlessly! In addition, it already has the new format of TypeError exception recently adopted to the remaining automaton types.

I have to incorporate into from_txt() the ability to return the object directly, instead of just returning the properties. I'll probably set up an attribute to allow me to decide among the two options and set the newer one (returning the object) as the default one.