.transitions(), .process_symbol(s)() implemented
The method .process_symbol() – which simply redirects to .process_symbols() – was made to overwrite the superclass' implementation. Why create two functions, you ask? Well, when processing states in a Pushdown Automaton, we are processing two symbols deep down: one for the word, one for the stack. As such, symbols with an "s".

The devil lies in the details.