.process_symbols() improved; .process_word() taking form
The method process_symbols() is evolving. It has two modes now: the first one is the default, and is the same as before; however, taking Eduardo's idea into consideration without radically changing any more code (at least not before getting this thing right), I developed a new mode that returns the whole stack instead of just the symbol to be written.