.process_word() further improved
Substantial progress has been made, with the method properly iterating through the states and performing the appropriate operations. The parts that are still missing are the additional behavior expected when dealing with ("q0", "?", "?") transitions, which essentially verify if everything is empty before moving to the final state.

To do that, I'll have to deal with .process_symbols() as well.