Dashed versions of Finite Pushdown Automata added
These versions were conceived to be tested against the loader. The goal here is to meet the Professor's requirements regarding the files that are going to be given to be loaded. Since the change is going to be made under the from_txt() function, however, the effect is going to be global. In other words: all automata classes and subclasses are going to have the same capabilities of loading with empty symbols (ε) or with dashes (-).

In addition, these changes will allow me to test the same automaton file against this Python implementation against the JavaScript one our group is building under tubskleyson/hephaestus, since there's being an issue with the empty transitions over there, or something of the kind. Let's see how this goes.