.process_symbols() and .process_word() entirely recreated
Given that there were quite some imprecise outputs for certain automata and words, an entire rework had to be made from the ground up. Fortunately, given all the previous experiences, this rework was quite staightforward.

The .process_symbols() took a lot of the heavy-lifting that was formerly done by .process_word(), and it now does its work in a cleaner way. With the new arrangement, less code is needed to get the job done, and debugging it is way less confusing as well.

In addition, the final verification is now done by the .process_word() instead of .process_symbols().

Finally, the printing statements are way more organized now. Weehoo!

I just need to update the documentation to reflect the parameter changes I made.